refactor claim extractor
