+++ src/NuGetGallery/Authentication/Providers/MicrosoftAccount/MicrosoftAccountAuthenticator.cs
@@ -1,8 +1,10 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System.Security.Claims;
 using System.Web.Mvc;
 using Microsoft.Owin.Security.MicrosoftAccount;
+using NuGetGallery.Authentication.Providers.Utils;
 using NuGetGallery.Configuration;
 using Owin;
 
@@ -36,6 +38,11 @@ namespace NuGetGallery.Authentication.Providers.MicrosoftAccount
         public override ActionResult Challenge(string redirectUrl)
         {
             return new ChallengeResult(BaseConfig.AuthenticationType, redirectUrl);
+        }
+
+        public override AuthInformation GetAuthInformation(ClaimsIdentity claimsIdentity)
+        {
+            return MSAIdentityExtractor.GetAuthInformation(claimsIdentity, DefaultAuthenticationType);
         }
     }
 }
